+++ packages/ricos-editor/web/src/utils/renderBlockComponent.tsx
@@ -1,6 +1,8 @@
 import type { FC } from 'react';
 import React from 'react';
 
+const offset = { x: 40 };
+
 export const renderSideBlockComponent = (
   id: string,
   dir: string,
@@ -8,18 +10,15 @@ export const renderSideBlockComponent = (
 ) => {
   const node = document.querySelectorAll(`[data-offset-key="${id}-0-0"]`)[0] as HTMLElement;
   if (node?.offsetParent) {
-    const top = node.getBoundingClientRect().top;
+    const { top, x } = node.getBoundingClientRect();
     const parentTop = node.offsetParent.getBoundingClientRect().top;
-    const lineHeight = window.getComputedStyle(node).getPropertyValue('line-height');
     return (
       <div
         dir={dir}
         style={{
           position: 'absolute',
-          display: 'block',
           top: top - parentTop,
-          lineHeight,
-          insetInlineEnd: 0,
+          insetInlineEnd: x - offset.x,
         }}
       >
         <Component id={id} />